activities-service/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
pymongo==4.6.1
pika==1.3.2
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2
PyJWT==2.8.0
requests==2.31.0


